[Norm] norminette에 맞게 수정

+++ 42cursus/philosophers/philo/time.c
@@ -6,7 +6,7 @@
 /*   By: yerilee <yerilee@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/21 19:14:33 by yerilee           #+#    #+#             */
-/*   Updated: 2023/12/21 19:14:40 by yerilee          ###   ########.fr       */
+/*   Updated: 2023/12/21 19:31:36 by yerilee          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,6 @@ long long	get_timestamp(void)
 	long long		result;
 
 	gettimeofday(&current_time, NULL);
-	result = ((size_t)current_time.tv_sec * 1000) + ((size_t)current_time.tv_usec / 1000);
+	result = (current_time.tv_sec * 1000) + (current_time.tv_usec / 1000);
 	return (result);
 }
